feat: navigate lessons sequentially
